# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The health status"
          }
        },
        "summary": "Get health status",
        "tags": [
          "Health"
        ]
      }
    },
    "/health/readiness": {
      "get": {
        "operationId": "HealthController_getReadiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The readiness status"
          }
        },
        "summary": "Get readiness status",
        "tags": [
          "Health"
        ]
      }
    },
    "/user": {
      "get": {
        "operationId": "UserController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all users",
        "tags": [
          "User"
        ]
      }
    },
    "/user/auth": {
      "get": {
        "operationId": "UserController_getUsersAuth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all users",
        "tags": [
          "User"
        ]
      }
    },
    "/services": {
      "get": {
        "operationId": "ServiceController_findAll",
        "parameters": [
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "description": "Filter by organization ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all services",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "ServiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDtoClass"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a new service",
        "tags": [
          "services"
        ]
      }
    },
    "/services/{id}": {
      "get": {
        "operationId": "ServiceController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Service not found"
          }
        },
        "summary": "Get service by ID",
        "tags": [
          "services"
        ]
      },
      "put": {
        "operationId": "ServiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDtoClass"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Service not found"
          }
        },
        "summary": "Update a service",
        "tags": [
          "services"
        ]
      }
    },
    "/services/organization/{organizationId}": {
      "get": {
        "operationId": "ServiceController_findByOrganizationId",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organization ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of services for the organization",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all services for an organization",
        "tags": [
          "services"
        ]
      }
    },
    "/services/{id}/soft": {
      "delete": {
        "operationId": "ServiceController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Service not found"
          }
        },
        "summary": "Soft delete a service (deactivate)",
        "tags": [
          "services"
        ]
      }
    },
    "/webhooks/clerk": {
      "post": {
        "operationId": "ClerkWebhookController_handleClerkWebhook",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ClerkWebhook"
        ]
      }
    }
  },
  "info": {
    "title": "Scheduler App API",
    "description": "Scheduler App API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "scheduler",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "UserResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The ID of the user"
          },
          "firstName": {
            "type": "object",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "object",
            "description": "The last name of the user"
          },
          "email": {
            "type": "string",
            "description": "The email of the user"
          },
          "status": {
            "type": "string",
            "description": "The status of the user"
          },
          "clerkId": {
            "type": "string",
            "description": "The clerk ID of the user"
          }
        },
        "required": [
          "id",
          "email",
          "status",
          "clerkId"
        ]
      },
      "ServiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Service ID"
          },
          "organizationId": {
            "type": "number",
            "description": "Organization ID"
          },
          "categoryId": {
            "type": "number",
            "description": "Category ID"
          },
          "name": {
            "type": "string",
            "description": "Service name"
          },
          "description": {
            "type": "string",
            "description": "Service description"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duration in minutes"
          },
          "paddingBeforeMinutes": {
            "type": "number",
            "description": "Padding before in minutes"
          },
          "paddingAfterMinutes": {
            "type": "number",
            "description": "Padding after in minutes"
          },
          "priceCents": {
            "type": "number",
            "description": "Price in cents"
          },
          "currency": {
            "type": "string",
            "description": "Currency code"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is service active"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created at timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at timestamp"
          }
        },
        "required": [
          "id",
          "organizationId",
          "name",
          "durationMinutes",
          "paddingBeforeMinutes",
          "paddingAfterMinutes",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateServiceDtoClass": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "number",
            "description": "Organization ID"
          },
          "categoryId": {
            "type": "number",
            "description": "Category ID"
          },
          "name": {
            "type": "string",
            "description": "Service name"
          },
          "description": {
            "type": "string",
            "description": "Service description"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duration in minutes"
          },
          "paddingBeforeMinutes": {
            "type": "number",
            "description": "Padding before in minutes",
            "default": 0
          },
          "paddingAfterMinutes": {
            "type": "number",
            "description": "Padding after in minutes",
            "default": 0
          },
          "priceCents": {
            "type": "number",
            "description": "Price in cents"
          },
          "currency": {
            "type": "string",
            "description": "Currency code (e.g., EUR, USD)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is service active",
            "default": true
          }
        },
        "required": [
          "organizationId",
          "name",
          "durationMinutes"
        ]
      },
      "UpdateServiceDtoClass": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "number",
            "description": "Category ID"
          },
          "name": {
            "type": "string",
            "description": "Service name"
          },
          "description": {
            "type": "string",
            "description": "Service description"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duration in minutes"
          },
          "paddingBeforeMinutes": {
            "type": "number",
            "description": "Padding before in minutes"
          },
          "paddingAfterMinutes": {
            "type": "number",
            "description": "Padding after in minutes"
          },
          "priceCents": {
            "type": "number",
            "description": "Price in cents"
          },
          "currency": {
            "type": "string",
            "description": "Currency code (e.g., EUR, USD)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is service active"
          }
        }
      }
    }
  }
}